Numerous minor fixes. bump installer version.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/branches/v2.3_dev@11891 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ phar_installer/app/lib/class.design_importer.php
@@ -114,7 +114,7 @@ class design_importer
 
         try {
             $tpl = \CmsLayoutTemplate::load( $tpl_name );
-            $tpl->delete();
+            if( $tpl ) $tpl->delete();
         }
         catch( \CmsDataNotFoundException $e ) {
             // not an error.
